Separate profile input fields and add previews
