+++ feature/login/src/main/java/com/epmedu/animeal/login/profile/ui/ProfileBirthDateInput.kt
@@ -1,92 +1,25 @@
 package com.epmedu.animeal.login.profile.ui
 
+import android.content.res.Configuration
 import androidx.compose.foundation.text.KeyboardActions
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.ui.focus.FocusDirection
 import androidx.compose.ui.focus.FocusManager
+import androidx.compose.ui.platform.LocalFocusManager
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.input.KeyboardType
+import androidx.compose.ui.tooling.preview.Preview
 import com.epmedu.animeal.extensions.DAY_MONTH_YEAR_DOT_FORMATTER
 import com.epmedu.animeal.extensions.formatDateToString
 import com.epmedu.animeal.foundation.dialog.DatePickerDialog
-import com.epmedu.animeal.foundation.input.TextInputField
+import com.epmedu.animeal.foundation.theme.AnimealTheme
 import com.epmedu.animeal.login.profile.presentation.FinishProfileEvent
 import com.epmedu.animeal.resources.R
 import com.vanpra.composematerialdialogs.rememberMaterialDialogState
 import java.time.LocalDate
-
-@Composable
-fun NameInput(
-    value: String,
-    error: String?,
-    focusManager: FocusManager,
-    isEnabled: Boolean = true,
-    onValueChange: (String) -> Unit,
-    onFocusRelease: () -> Unit,
-) {
-    TextInputField(
-        isEnabled = isEnabled,
-        title = stringResource(id = R.string.profile_name_title),
-        hint = stringResource(id = R.string.profile_name_hint),
-        onValueChange = onValueChange,
-        value = value,
-        errorText = error,
-        onClearFocus = onFocusRelease,
-        keyboardOptions = KeyboardOptions(imeAction = ImeAction.Next),
-        keyboardActions = KeyboardActions(onNext = { focusManager.moveFocus(FocusDirection.Down) })
-    )
-}
-
-@Composable
-fun SurnameInput(
-    value: String,
-    error: String?,
-    focusManager: FocusManager,
-    isEnabled: Boolean = true,
-    onValueChange: (String) -> Unit,
-    onFocusRelease: () -> Unit,
-) {
-    TextInputField(
-        isEnabled = isEnabled,
-        title = stringResource(id = R.string.profile_surname_title),
-        hint = stringResource(id = R.string.profile_surname_hint),
-        onValueChange = onValueChange,
-        value = value,
-        errorText = error,
-        onClearFocus = onFocusRelease,
-        keyboardOptions = KeyboardOptions(imeAction = ImeAction.Next),
-        keyboardActions = KeyboardActions(onNext = { focusManager.moveFocus(FocusDirection.Down) })
-    )
-}
-
-@Composable
-fun EmailInput(
-    value: String,
-    error: String?,
-    focusManager: FocusManager,
-    isEnabled: Boolean = true,
-    onValueChange: (String) -> Unit,
-    onFocusRelease: () -> Unit,
-) {
-    TextInputField(
-        isEnabled = isEnabled,
-        title = stringResource(id = R.string.profile_email_title),
-        hint = stringResource(id = R.string.profile_email_hint),
-        onValueChange = onValueChange,
-        value = value,
-        errorText = error,
-        onClearFocus = onFocusRelease,
-        keyboardOptions = KeyboardOptions(
-            imeAction = ImeAction.Done,
-            keyboardType = KeyboardType.Email
-        ),
-        keyboardActions = KeyboardActions(onDone = { focusManager.clearFocus() })
-    )
-}
 
 @Suppress("LongParameterList")
 @Composable
@@ -138,4 +71,20 @@ fun BirthDateInput(
             showDialog.value = false
         }
     )
+}
+
+@Preview
+@Preview(uiMode = Configuration.UI_MODE_NIGHT_YES)
+@Composable
+private fun BirthDateInputPreview() {
+    AnimealTheme {
+        BirthDateInput(
+            value = "11.11.1900",
+            error = null,
+            initialDate = LocalDate.now(),
+            focusManager = LocalFocusManager.current,
+            onValueChange = {},
+            onFocusRelease = {}
+        )
+    }
 }
